Move tests out of script
[skip ci]
